orderdesk/fix: TableBoard에서는 order-summary API 데이터만 사용하도록 캐시 전략 변경

+++ apps/orderdesk/src/app/(navigator)/(sidebar)/stores/[storeId]/orders/components/table-order-detail/columns.tsx
@@ -82,7 +82,7 @@ export const columns: ColumnDef<ResponseOrderItemWithOrderId>[] = [
     accessorKey: "totalPrice",
     header: () => <div className="text-right w-full">가격</div>,
     cell: ({ row }) => {
-      const amount = parseFloat(row.getValue("totalPrice"));
+      const amount = parseInt(row.getValue("totalPrice"));
       const formatted = new Intl.NumberFormat("ko-KR", {
         currency: "KRW",
       }).format(amount);
